+++ packages/layout/hooks/useAppMessage.ts
@@ -1,0 +1,6 @@
+import { App } from 'antd'
+
+export function useAppMessage() {
+  const { message, modal } = App.useApp()
+  return { message, modal }
+}